+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+
+## [0.1.16] - 2025-12-08
+
+### Added
+- NIfTI file compression functionality (`nii_to_niigz`)
+- NIfTI file writing functionality (`writeNifti`) (also via `writeVTKFile` for nii and nii.gz)
+
+
+### Changed
+- Updated documentation
 ## [0.1.15] - 2025-11-25
 
 ### Fixed